Move `log.hpp` from `backend` to `detail`.
