+++ include/spblas/detail/log.hpp
